:wrench: uses one index.js for AMD and CJS (made possible because the parser is in a module loader now as well)
